requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
install==1.3.5
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
psycopg2-binary==2.9.3